refactor: common metadata
